git-svn-id: http://yagf.googlecode.com/svn/trunk@246 bb55e26e-657b-11de-982f-55b50260c994

+++ src/qsnippet.cpp
@@ -47,7 +47,7 @@ void QSnippet::setRotation(qreal rot)
     rotation = rot;
 }
 
-int QSnippet::getRotation()
+qreal QSnippet::getRotation()
 {
     return rotation;
 }
